[fix,settings] 폴더링 작업 및 onboarding flow 진행중 뒤로가기,다음 이동시 복원 안되는 현상 수정
